fix: move IntoSymbols trait to symbol module
