+++ commons/src/scanner/symbol.rs
@@ -2,7 +2,10 @@
 
 use core::fmt;
 
-use super::position::{Offset, Position};
+use super::{
+    grapheme_split,
+    position::{Offset, Position},
+};
 
 /// Possible kinds of Symbol found in Unimarkup document.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
@@ -228,5 +231,45 @@ impl From<&str> for SymbolKind {
             }
             _ => SymbolKind::Plain,
         }
+    }
+}
+
+/// Trait for conversion of input into Unimarkup symbols.
+pub trait IntoSymbols<'s> {
+    type Output: AsRef<[Symbol<'s>]>;
+
+    /// Converts input into collection of Unimarkup symbols.
+    fn into_symbols(self) -> Self::Output;
+}
+
+impl<'s> IntoSymbols<'s> for &'s str {
+    type Output = Vec<Symbol<'s>>;
+
+    fn into_symbols(self) -> Self::Output {
+        grapheme_split(self)
+    }
+}
+
+impl<'s> IntoSymbols<'s> for Vec<Symbol<'s>> {
+    type Output = Vec<Symbol<'s>>;
+
+    fn into_symbols(self) -> Self::Output {
+        self
+    }
+}
+
+impl<'s> IntoSymbols<'s> for &'s Vec<Symbol<'s>> {
+    type Output = &'s [Symbol<'s>];
+
+    fn into_symbols(self) -> Self::Output {
+        self
+    }
+}
+
+impl<'s> IntoSymbols<'s> for &'s [Symbol<'s>] {
+    type Output = &'s [Symbol<'s>];
+
+    fn into_symbols(self) -> Self::Output {
+        self
     }
 }
